third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/fixed-layout-excess-width-distribution-001-expected.txt

This is a testharness.js-based test.
FAIL #theTable 1 assert_equals: 
<table id="theTable">
  <tbody><tr>
    <td data-expected-width="180"></td>
    <td data-expected-width="90"></td>
    <td data-expected-width="30"></td>
  </tr>
</tbody></table>
width expected 180 but got 100
Harness: the test ran to completion.
